integrate with typeorm

+++ src/typeorm/service/GroupMenuService.ts
@@ -14,24 +14,32 @@ export class GroupMenuService {
         const founds = await this.repo.findBy({ groupChat: { jid } })
         return founds
     }
+    // async findAllTextMenu(jid: string): Promise<GroupMenuEntity[]> {
+    //     const founds = await this.repo.findBy({ groupChat: { jid }, type: GroupMenuType.TEXT })
+    //     return founds
+    // }
+    // async findAllImageMenu(jid: string): Promise<GroupMenuEntity[]> {
+    //     const founds = await this.repo.findBy({ groupChat: { jid }, type: GroupMenuType.IMAGE })
+    //     return founds
+    // }
 
     async findOneMenu(jid: string, key: string) {
         const found = await this.repo.findOneBy({ groupChat: { jid }, key })
-        if (!found) {
-            throw new Error(`${key} not found in ${jid}`)
-        }
         return found
     }
 
     async createMenu(groupChat: GroupChatEntity, key: string, value: string, type: GroupMenuType) {
-        const groupMenu = this.repo.create({
-            groupChat,
-            key,
-            value,
-            type
-        })
-
-        return await this.repo.save(groupMenu)
+        const found = await this.findOneMenu(groupChat.jid, key)
+        if (!found){
+            const groupMenu = this.repo.create({
+                groupChat,
+                key,
+                value,
+                type
+            })
+    
+            return await this.repo.save(groupMenu)
+        }
     }
     async createMenuText(groupChat: GroupChatEntity, key: string, value: string) {
         return this.createMenu(groupChat, key, value, GroupMenuType.TEXT)
@@ -42,8 +50,10 @@ export class GroupMenuService {
 
     async updateMenuValue(jid: string, key: string, value: string) {
         const found = await this.findOneMenu(jid, key)
-        found.value = value
-        return await this.repo.save(found)
+        if(found){
+            found.value = value
+            return await this.repo.save(found)
+        }
     }
 
     async removeAllMenu(jid: string) {
@@ -53,7 +63,9 @@ export class GroupMenuService {
 
     async removeOneMenu(jid: string, key: string) {
         const found = await this.findOneMenu(jid, key)
-        return this.repo.remove(found)
+        if (found){
+            return this.repo.remove(found)
+        }
     }
 
 }
